zonotope.*, parallelotope.*:   Put back vertices() methods ( where they belong :) )

+++ include/geometry/grid.h
@@ -119,7 +119,7 @@ namespace Ariadne {
       /*! \brief The coordinate of the \a n th subdivision point in dimension \a d. */
       virtual real_type subdivision_coordinate(dimension_type d, index_type n) const {
         assert(d<this->dimension());
-        assert(0<=n && n<_subdivision_coordinates[d].size());
+        assert(0<=n && uint(n)<_subdivision_coordinates[d].size());
         return _subdivision_coordinates[d][n];
       }
